refactor: move binder package to pkg directory
Relocate binder package from root to pkg/binder/ as part of
framework reorganization. No functional changes.
