+++ pkg/binder/path_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dmitrymomot/saaskit/binder"
+	"github.com/dmitrymomot/saaskit/pkg/binder"
 )
 
 func TestPath(t *testing.T) {
