+++ pkg/binder/realworld_saas_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dmitrymomot/saaskit/binder"
+	"github.com/dmitrymomot/saaskit/pkg/binder"
 )
 
 func TestRealWorldSaaSQueryScenarios(t *testing.T) {
